+++ syn/base/b/meta.py
@@ -9,7 +9,7 @@ from functools import partial
 from syn.base.a.meta import Attr as _Attr
 from syn.base.a.meta import Attrs as _Attrs
 from syn.base.a.meta import Meta as _Meta
-from syn.base.a.meta import combine
+from syn.base.a.meta import combine, Data
 
 _OAttr = partial(_Attr, optional=True)
 
@@ -111,6 +111,6 @@ class Meta(_Meta):
 #-------------------------------------------------------------------------------
 # __all__
 
-__all__ = ('Attr', 'Attrs', 'Meta')
+__all__ = ('Attr', 'Attrs', 'Meta', 'Data')
 
 #-------------------------------------------------------------------------------